requirements.txt
altair
numpy
pandas
pydeck
streamlit
pika
python-dotenv